requirements.txt
PyQt6>=6.0.0
opencv-python>=4.5.0
numpy>=1.19.0
ultralytics>=8.0.0
Pillow>=8.0.0
matplotlib>=3.3.0
pydantic>=2.0.0 
